rename all files with mocks to end with consistent `*.mocks.*` rather than with `*.mock(s).*`
