+++ jsapp/js/components/activity/FormActivity.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
 import { reactRouterParameters, withRouter } from 'storybook-addon-remix-react-router'
 import { expect, userEvent, waitFor, within } from 'storybook/test'
-import assetHistoryMock, { mockAssetUid } from '#/endpoints/assetHistory.mock'
-import assetHistoryActionsMock from '#/endpoints/assetHistoryActions.mock'
-import { queryClientDecorator } from '#/query/queryClient.mock'
+import assetHistoryMock, { mockAssetUid } from '#/endpoints/assetHistory.mocks'
+import assetHistoryActionsMock from '#/endpoints/assetHistoryActions.mocks'
+import { queryClientDecorator } from '#/query/queryClient.mocks'
 import { ROUTES } from '#/router/routerConstants'
 import FormActivity from './FormActivity'
