+++ jsapp/js/endpoints/assetHistory.mocks.ts
