+++ jsapp/js/endpoints/assetHistoryActions.mocks.ts
@@ -1,7 +1,7 @@
 import { http, HttpResponse, type PathParams } from 'msw'
 import { endpoints } from '#/api.endpoints'
 import type { AssetHistoryActionsResponse } from '#/components/activity/activity.constants'
-import { mockAssetUid } from './assetHistory.mock'
+import { mockAssetUid } from './assetHistory.mocks'
 
 /**
  * Mock API for listing all available history actions. Use it in Storybook tests in `parameters.msw.handlers[]`.
